Feature/contract init (#25)
* Create timelock for public sale

* Add initializer with mint

* fix merge issue
