+++ contracts/token/RarePizzasBox.sol
@@ -18,8 +18,8 @@ import "../interfaces/IRarePizzasBox.sol";
  * @dev Rare Pizzas Box mints pizza box token for callers who call the purchase function.
  */
 contract RarePizzasBox is
-    ERC721EnumerableUpgradeable,
     OwnableUpgradeable,
+    ERC721EnumerableUpgradeable,
     BondingCurve,
     IRarePizzasBox,
     IOpenSeaCompatible
@@ -29,18 +29,30 @@ contract RarePizzasBox is
     using CountersUpgradeable for CountersUpgradeable.Counter;
     using SafeMathUpgradeable for uint256;
 
-    uint256 private constant MAX_TOKEN_SUPPLY = 10000;
-    uint256 private constant MAX_MINTABLE_SUPPLY = 1250; // TODO: verify
+    // V1 Variables (do not modify this section when upgrading)
 
-    CountersUpgradeable.Counter private _minted_pizza_count;
-    CountersUpgradeable.Counter private _whole_pizza_count;
+    // 2021-03-14:15h::9m::26s
+    uint256 public _public_sale_start_timestamp = 1615734566;
 
-    // TODO:
-    string private _boxMetadataUri =
+    uint256 public constant MAX_TOKEN_SUPPLY = 10000;
+    uint256 public constant MAX_MINTABLE_SUPPLY = 1250;
+    uint256 public constant MAX_PURCHASABLE_SUPPLY = 8750;
+
+    string public _boxMetadataUri =
         "https://ipfs.io/ipfs/some/path/to/box/metadata";
 
-    function initialize() public initializer {
+    CountersUpgradeable.Counter public _minted_pizza_count;
+    CountersUpgradeable.Counter public _purchased_pizza_count;
+
+    // END V1 Variables
+
+    function initialize(address mintTo, uint8 count) public initializer {
+        __Ownable_init();
         __ERC721_init("Rare Pizza Box", "RAREPIZZASBOX");
+
+        if (count > 0) {
+            mint(mintTo, count);
+        }
     }
 
     // IOpenSeaCompatible
@@ -58,20 +70,13 @@ contract RarePizzasBox is
         return MAX_TOKEN_SUPPLY;
     }
 
-    function purchase() public payable override {
-        require(
-            totalSupply() <= maxSupply(),
-            "RAREPIZZA: Call exceeds maximum supply"
-        );
-
-        require(
-            totalSupply().add(1) <= maxSupply(),
-            "RAREPIZZA: purchase would exceed maxSupply"
-        );
+    function purchase() public payable virtual override {
+        require(block.timestamp >= _public_sale_start_timestamp, "RAREPIZZA: The sale hasn't started yet");
+        require(totalSupply().add(1) <= MAX_TOKEN_SUPPLY, 'RAREPIZZA: purchase would exceed maxSupply');
 
         uint256 price = getPrice();
-        require(price == msg.value, "RAREPIZZA: price must be on the curve");
-        _whole_pizza_count.increment();
+        require(price == msg.value, 'RAREPIZZA: price must be on the curve');
+        _purchased_pizza_count.increment();
         _safeMint(msg.sender, _getNextPizzaTokenId());
     }
 
@@ -90,35 +95,27 @@ contract RarePizzasBox is
         return string(abi.encodePacked(_boxMetadataUri));
     }
 
-    /**
-     * @dev See {IERC721Enumerable-totalSupply}.
-     */
-    function totalSupply()
-        public
-        view
-        override(ERC721EnumerableUpgradeable, IERC721EnumerableUpgradeable)
-        returns (uint256)
-    {
-        return _whole_pizza_count.current();
-    }
-
     // Member Functions
 
     /**
      * allows the contract owner to mint up to a specific number of boxes
      */
-    function mint(address to) public onlyOwner {
+    function mint(address to, uint8 count) public onlyOwner {
+        require(count > 0, "RAREPIZZA: must provide a count");
+
         require(
-            totalSupply().add(1) <= maxSupply(),
+            totalSupply().add(count) <= maxSupply(),
             "RAREPIZZA: mint would exceed maxSupply"
         );
         require(
-            _minted_pizza_count.current().add(1) <= MAX_MINTABLE_SUPPLY,
+            _minted_pizza_count.current().add(count) <= MAX_MINTABLE_SUPPLY,
             "RAREPIZZA: mint would exceed MAX_MINTABLE_SUPPLY"
         );
-        _minted_pizza_count.increment();
-        _whole_pizza_count.increment();
-        _safeMint(to, _getNextPizzaTokenId());
+
+        for (uint256 i = 0; i < count; i++) {
+            _minted_pizza_count.increment();
+            _safeMint(to, _getNextPizzaTokenId());
+        }
     }
 
     function _getNextPizzaTokenId() private view returns (uint256) {
